drivers: dac: ad3552r_iio: change iio_ad3552r_attr_get len type
iio_ad3552r_attr_get() len type changed to match the struct
iio_attribute show and store prototypes.

+++ drivers/dac/ad3552r/iio_ad3552r.c
@@ -82,7 +82,7 @@ struct iio_ad3552r_desc {
 /************************* Functions Definitions *****************************/
 /*****************************************************************************/
 
-static int iio_ad3552r_attr_get(void *device, char *buf, size_t len,
+static int iio_ad3552r_attr_get(void *device, char *buf, uint32_t len,
 				const struct iio_ch_info *channel, intptr_t priv)
 {
 	struct iio_ad3552r_desc *iio_dac = device;
